remove networkManager
