+++ lib/msal-common/src/client/BaseClient.ts
@@ -8,8 +8,8 @@ import {
     buildClientConfiguration,
     CommonClientConfiguration,
 } from "../config/ClientConfiguration.js";
-import { INetworkModule } from "../network/INetworkModule.js";
-import { NetworkManager, NetworkResponse } from "../network/NetworkManager.js";
+import { INetworkModule, NetworkRequestOptions } from "../network/INetworkModule.js";
+import { NetworkResponse } from "../network/NetworkResponse.js";
 import { ICrypto } from "../crypto/ICrypto.js";
 import { Authority } from "../authority/Authority.js";
 import { Logger } from "../logger/Logger.js";
@@ -26,6 +26,9 @@ import { RequestParameterBuilder } from "../request/RequestParameterBuilder.js";
 import { BaseAuthRequest } from "../request/BaseAuthRequest.js";
 import { createDiscoveredInstance } from "../authority/AuthorityFactory.js";
 import { PerformanceEvents } from "../telemetry/performance/PerformanceEvent.js";
+import { ThrottlingUtils } from "../network/ThrottlingUtils.js";
+import { AuthError } from "../error/AuthError.js";
+import { ClientAuthErrorCodes, createClientAuthError } from "../error/ClientAuthError.js";
 
 /**
  * Base application class which will construct requests to send to and handle responses from the Microsoft STS using the authorization code flow.
@@ -49,9 +52,6 @@ export abstract class BaseClient {
 
     // Server Telemetry Manager
     protected serverTelemetryManager: ServerTelemetryManager | null;
-
-    // Network Manager
-    protected networkManager: NetworkManager;
 
     // Default authority object
     public authority: Authority;
@@ -77,12 +77,6 @@ export abstract class BaseClient {
 
         // Set the network interface
         this.networkClient = this.config.networkInterface;
-
-        // Set the NetworkManager
-        this.networkManager = new NetworkManager(
-            this.networkClient,
-            this.cacheManager
-        );
 
         // Set TelemetryManager
         this.serverTelemetryManager = this.config.serverTelemetryManager;
@@ -152,19 +146,12 @@ export abstract class BaseClient {
         }
 
         const response =
-            await this.networkManager.sendPostRequest<ServerAuthorizationTokenResponse>(
+            await this.sendPostRequest<ServerAuthorizationTokenResponse>(
                 thumbprint,
                 tokenEndpoint,
-                { body: queryString, headers: headers }
+                { body: queryString, headers: headers },
+                correlationId
             );
-        this.performanceClient?.addFields(
-            {
-                refreshTokenSize: response.body.refresh_token?.length || 0,
-                httpVerToken:
-                    response.headers?.[HeaderNames.X_MS_HTTP_VERSION] || "",
-            },
-            correlationId
-        );
 
         if (
             this.config.serverTelemetryManager &&
@@ -174,6 +161,50 @@ export abstract class BaseClient {
             // Telemetry data successfully logged by server, clear Telemetry cache
             this.config.serverTelemetryManager.clearTelemetryCache();
         }
+
+        return response;
+    }
+
+    /**
+     * Wraps sendPostRequestAsync with necessary preflight and postflight logic
+     * @param thumbprint
+     * @param tokenEndpoint
+     * @param options
+     * @param correlationId
+     */
+    async sendPostRequest<T extends ServerAuthorizationTokenResponse>(
+        thumbprint: RequestThumbprint,
+        tokenEndpoint: string,
+        options: NetworkRequestOptions,
+        correlationId: string
+    ): Promise<NetworkResponse<T>> {
+        ThrottlingUtils.preProcess(this.cacheManager, thumbprint);
+
+        let response;
+        try {
+            response = await this.networkClient.sendPostRequestAsync<T>(
+                tokenEndpoint,
+                options
+            );
+            const responseHeaders = response.headers || {};
+            this.performanceClient?.addFields(
+                {
+                    refreshTokenSize: response.body.refresh_token?.length || 0,
+                    httpVerToken:
+                        responseHeaders[HeaderNames.X_MS_HTTP_VERSION] || "",
+                    requestId: responseHeaders[HeaderNames.X_MS_REQUEST_ID] || ""
+                },
+                correlationId
+            );
+        } catch (e) {
+            if (e instanceof AuthError) {
+                throw e;
+            } else {
+                throw createClientAuthError(ClientAuthErrorCodes.networkError);
+            }
+        }
+
+        ThrottlingUtils.postProcess(this.cacheManager, thumbprint, response);
 
         return response;
     }
